third_party/blink/web_tests/platform/mac/external/wpt/css/css-sizing/keyword-sizes-for-intrinsic-contributions.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .test 1
  assert_equals: \n<div class="test width" data-expected-width="70"><div style="width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 2
  assert_equals: \n<div class="test width" data-expected-width="70"><div style="width: fit-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 3
  assert_equals: \n<div class="test width" data-expected-width="70"><div class="stretch">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 4
  assert_equals: \n<div class="test width" data-expected-width="150"><div style="width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 5
  assert_equals: \n<div class="test width" data-expected-width="70"><div style="width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 8
  assert_equals: \n<div class="test width" data-expected-width="150"><div style="width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 9
  assert_equals: \n<div class="test width" data-expected-width="70"><div style="width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 10
  assert_equals: \n<div class="test width" data-expected-width="150"><div style="width: fit-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 11
  assert_equals: \n<div class="test width" data-expected-width="150"><div class="stretch">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 12
  assert_equals: \n<div class="test width" data-expected-width="150"><div style="width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 13
  assert_equals: \n<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 14
  assert_equals: \n<div class="test min-width" data-expected-width="70"><div style="min-width: fit-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 15
  assert_equals: \n<div class="test min-width" data-expected-width="70"><div class="stretch">XXX XXX</div></div>\nwidth expected 70 but got 10
[FAIL] .test 16
  assert_equals: \n<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 17
  assert_equals: \n<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 18
  assert_equals: \n<div class="test min-width" data-expected-width="90"><div style="min-width: fit-content">XXX XXX</div></div>\nwidth expected 90 but got 53
[FAIL] .test 19
  assert_equals: \n<div class="test min-width" data-expected-width="90"><div class="stretch">XXX XXX</div></div>\nwidth expected 90 but got 10
[FAIL] .test 20
  assert_equals: \n<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 21
  assert_equals: \n<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 22
  assert_equals: \n<div class="test min-width" data-expected-width="150"><div style="min-width: fit-content">XXX XXX</div></div>\nwidth expected 150 but got 53
[FAIL] .test 23
  assert_equals: \n<div class="test min-width" data-expected-width="150"><div class="stretch">XXX XXX</div></div>\nwidth expected 150 but got 10
[FAIL] .test 24
  assert_equals: \n<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 25
  assert_equals: \n<div class="test max-width" data-expected-width="70"><div style="max-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 26
  assert_equals: \n<div class="test max-width" data-expected-width="70"><div style="max-width: fit-content">XXX XXX</div></div>\nwidth expected 70 but got 102
[FAIL] .test 27
  assert_equals: \n<div class="test max-width" data-expected-width="70"><div class="stretch">XXX XXX</div></div>\nwidth expected 70 but got 210
[FAIL] .test 28
  assert_equals: \n<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 29
  assert_equals: \n<div class="test max-width" data-expected-width="70"><div style="max-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 30
  assert_equals: \n<div class="test max-width" data-expected-width="90"><div style="max-width: fit-content">XXX XXX</div></div>\nwidth expected 90 but got 102
[FAIL] .test 31
  assert_equals: \n<div class="test max-width" data-expected-width="90"><div class="stretch">XXX XXX</div></div>\nwidth expected 90 but got 210
[FAIL] .test 32
  assert_equals: \n<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 33
  assert_equals: \n<div class="test max-width" data-expected-width="70"><div style="max-width: min-content">XXX XXX</div></div>\nwidth expected 70 but got 53
[FAIL] .test 34
  assert_equals: \n<div class="test max-width" data-expected-width="150"><div style="max-width: fit-content">XXX XXX</div></div>\nwidth expected 150 but got 102
[FAIL] .test 35
  assert_equals: \n<div class="test max-width" data-expected-width="150"><div class="stretch">XXX XXX</div></div>\nwidth expected 150 but got 210
[FAIL] .test 36
  assert_equals: \n<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>\nwidth expected 150 but got 102
Harness: the test ran to completion.
